+++ src/app/(default)/new/_components/post-form.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { Controller, useForm } from 'react-hook-form'
 import { TextField } from '~/app/_components/text-field'
 import { Button } from '~/app/_components/button'
-import MarkdownIcon from '~/app/_svg/markdown-icon'
+import MarkdownIcon from '~/components/svg/markdown-icon'
 import { MarkdownEditor } from '~/app/_components/markdown-editor'
 import { useRouter } from 'next/navigation'
 import { api } from '~/trpc/react'
