+++ src/components/alert-dialog.tsx
@@ -1,18 +1,18 @@
 'use client'
 
-import * as DialogPrimitive from '@radix-ui/react-dialog'
+import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog'
 import * as React from 'react'
-import XIcon from '../_svg/x-icon'
+import XIcon from '~/components/svg/x-icon'
 import { useDialogStore } from '../_hooks/use-dialog-store'
 import { type ReactNode } from 'react'
 import { classNames } from '~/utils/core'
 
-export const Dialog = () => {
+export const AlertDialog = () => {
   const { isOpen, handleDialogClose, dialogContent } = useDialogStore()
 
   return (
-    <DialogPrimitive.Root open={isOpen} onOpenChange={handleDialogClose}>
-      <DialogPrimitive.Content
+    <AlertDialogPrimitive.Root open={isOpen} onOpenChange={handleDialogClose}>
+      <AlertDialogPrimitive.Content
         className={classNames(
           'fixed inset-0 z-10 overflow-y-auto',
           'radix-state-open:animate-scale-in-content',
@@ -20,34 +20,50 @@ export const Dialog = () => {
         )}
       >
         <div className="min-h-screen px-4 text-center">
-          <DialogPrimitive.Overlay className="fixed inset-0 transition-opacity bg-gray-700 opacity-90 dark:bg-gray-900" />
+          <AlertDialogPrimitive.Overlay className="fixed inset-0 transition-opacity bg-gray-700 opacity-90 dark:bg-gray-900" />
 
           <div className="inline-block w-full max-w-md mt-[15vh] mb-8 overflow-hidden text-left align-middle transition-all transform bg-primary rounded-lg shadow-xl dark:border">
             {dialogContent}
           </div>
         </div>
-      </DialogPrimitive.Content>
-    </DialogPrimitive.Root>
+      </AlertDialogPrimitive.Content>
+    </AlertDialogPrimitive.Root>
   )
 }
 
-export const DialogContent = ({ children }: { children: ReactNode }) => {
+export const AlertDialogContent = ({ children }: { children: ReactNode }) => {
   return <div className="px-6 pt-6 pb-12">{children}</div>
 }
 
-export const DialogActions = ({ children }: { children: ReactNode }) => {
+export const AlertDialogActions = ({ children }: { children: ReactNode }) => {
   return <div className="flex gap-4 px-6 py-4 border-t">{children}</div>
 }
 
-export const DialogTitle = ({ children }: { children: ReactNode }) => {
+export const AlertDialogAction = ({ children }: { children: ReactNode }) => {
   return (
-    <DialogPrimitive.Title asChild className="text-lg font-semibold">
-      <h3>{children}</h3>
-    </DialogPrimitive.Title>
+    <AlertDialogPrimitive.Action asChild>
+      {children}
+    </AlertDialogPrimitive.Action>
   )
 }
 
-export const DialogDescription = ({
+export const AlertDialogCancel = ({ children }: { children: ReactNode }) => {
+  return (
+    <AlertDialogPrimitive.Cancel asChild>
+      {children}
+    </AlertDialogPrimitive.Cancel>
+  )
+}
+
+export const AlertDialogTitle = ({ children }: { children: ReactNode }) => {
+  return (
+    <AlertDialogPrimitive.Title asChild className="text-lg font-semibold">
+      <h3>{children}</h3>
+    </AlertDialogPrimitive.Title>
+  )
+}
+
+export const AlertDialogDescription = ({
   children,
   className,
 }: {
@@ -55,13 +71,17 @@ export const DialogDescription = ({
   className?: string
 }) => {
   return (
-    <DialogPrimitive.Description className={className}>
+    <AlertDialogPrimitive.Description className={className}>
       {children}
-    </DialogPrimitive.Description>
+    </AlertDialogPrimitive.Description>
   )
 }
 
-export function DialogCloseButton({ onClick }: { onClick: () => void }) {
+export const AlertDialogCloseButton = ({
+  onClick,
+}: {
+  onClick: () => void
+}) => {
   return (
     <div className="absolute top-0 right-0 pt-6 pr-6">
       <button
