+++ src/components/banner.tsx
