+++ src/components/button.tsx
@@ -9,7 +9,7 @@ import { cva, type VariantProps } from 'class-variance-authority'
 import Link, { type LinkProps } from 'next/link'
 
 import { classNames } from '~/utils/core'
-import Spinner from '../_svg/spinner'
+import Spinner from '~/components/svg/spinner'
 
 const button = cva(
   'inline-flex items-center justify-center font-semibold transition-colors rounded-full focus-ring px-4 text-sm h-8',
