+++ src/components/post-action.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from 'next/navigation'
 import { toast } from 'react-hot-toast'
-import { Button } from '~/app/_components/button'
+import { Button } from '~/components/button'
 import {
   AlertDialogAction,
   AlertDialogActions,
@@ -11,7 +11,7 @@ import {
   AlertDialogDescription,
   AlertDialogTitle,
   AlertDialogCancel,
-} from '~/app/_components/alert-dialog'
+} from '~/components/alert-dialog'
 import {
   Menu,
   MenuButton,
@@ -20,12 +20,12 @@ import {
   MenuItem,
   MenuItemsContent,
   MenuItemLink,
-} from '~/app/_components/menu'
-import { useDialogStore } from '~/app/_hooks/use-dialog-store'
-import DotsIcon from '~/app/_svg/dots-icon'
-import EditIcon from '~/app/_svg/edit-icon'
-import EyeIcon from '~/app/_svg/eye-icon'
-import TrashIcon from '~/app/_svg/trash-icon'
+} from '~/components/menu'
+import { useDialogStore } from '~/hooks/use-dialog-store'
+import DotsIcon from '~/components/svg/dots-icon'
+import EditIcon from '~/components/svg/edit-icon'
+import EyeIcon from '~/components/svg/eye-icon'
+import TrashIcon from '~/components/svg/trash-icon'
 import { api } from '~/trpc/react'
 
 type PostActionProps = {
