+++ src/components/post-summary.tsx
@@ -6,8 +6,8 @@ import { classNames } from '~/utils/core'
 import { formatDistanceToNow } from 'date-fns/formatDistanceToNow'
 import { AuthorWithDate } from '~/app/_components/author-with-date'
 import { HtmlView } from '~/app/_components/html-view'
-import ChevronRightIcon from '~/app/_svg/chevron-right-icon'
-import MessageIcon from '~/app/_svg/message-icon'
+import ChevronRightIcon from '~/components/svg/chevron-right-icon'
+import MessageIcon from '~/components/svg/message-icon'
 
 import { ReactionButton } from './reaction-button'
 import { Suspense } from 'react'
