+++ src/components/svg/bold-icon.tsx
