+++ src/components/svg/chevron-left-icon.tsx
