+++ src/components/svg/dot-pattern.tsx
