+++ src/components/svg/edit-icon.tsx
