+++ src/components/svg/eye-closed-icon.tsx
