+++ src/components/svg/eye-icon.tsx
