+++ src/components/svg/italic-icon.tsx
