+++ src/components/svg/message-icon.tsx
