+++ src/components/svg/search-icon.tsx
