+++ src/components/svg/trash-icon.tsx
